// app/calendar/model/appoiment.ts
export class Appoiment {
    Id: number;
    ClinicId: number;
    EmployeeId: number;
    Status: string;
    Active: boolean;
    PatientId: number;
    PatientName: string;
    StartDate: string;
    EndDate: string;
    ClassName: string;
    Hour: string;
    strStartDate: string;
    strHour: string;
    Duration: number;
    CantDateType: number;
    DateType: string;
}